+++ assets/src/js/docs/ui-components/header.js
@@ -14,21 +14,3 @@ if (signedInDemoEl) {
     $.first('[data-auth-wrapper]', signedInDemoEl).classList.remove('is-hidden');
     $.first('[data-login]', signedInDemoEl).remove();
 }
-
-const demoLogos = $('.demo .c-logo');
-demoLogos.forEach(logo => {
-    logo.addEventListener('click', e => {
-        e.preventDefault();
-    });
-});
-
-const navEls = $('[data-navglobal]');
-navEls.forEach(nav => {
-    const navAnchors = $('a', nav);
-
-    navAnchors.forEach(anchor => {
-        anchor.addEventListener('click', e => {
-            e.preventDefault();
-        });
-    });
-});